+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+##TBD
+* Add SSO Seeding call in MSAL Test MacApp
+* Fix custom webview bug in MSAL Test MacApp
+*
 ## [1.1.8] - 2020-08-24
 * Disabling check for validating result Account.
 * Fix unused parameter errors and add macOS specific test mocks.